Fix similarity search failing for records without a title
titleMatchTier rendered its no-title branch as a bare (0), which Postgres
reads in ORDER BY as column position 0 and rejects, so search.byRecordId
returned 500 for any seed record with an empty title. Cast the constant
so it parses as an expression.

+++ src/server/lib/constants.ts
@@ -74,7 +74,8 @@ export const normalizedTitleColumn = (column: Column): SQL =>
  */
 export const titleMatchTier = (column: Column, normalizedTitle: string | null): SQL<number> =>
   normalizedTitle === null
-    ? sql<number>`(0)`
+    ? // The cast keeps ORDER BY from reading the bare constant as a column position.
+      sql<number>`(0::int)`
     : sql<number>`CASE
         WHEN ${normalizedTitleColumn(column)} = ${normalizedTitle} THEN 2
         WHEN similarity(${column}, ${normalizedTitle}) >= ${TITLE_VARIANT_SIMILARITY} THEN 1
